% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{AdaLasso}
\alias{AdaLasso}
\title{Extract DFC with Adaptive Lasso}
\usage{
AdaLasso(
  data,
  label,
  gamma = 1,
  lambda_penalty = "1se",
  lambda_weight = "1se",
  seed = NULL
)
}
\arguments{
\item{data}{scaled count gene x cell matrix}

\item{label}{logical or binary vector labeling target clusters}

\item{gamma}{Parameter to control the effect of penalty}

\item{lambda_penalty}{Parameter to control which lambda is used to calculate
penalty}

\item{lambda_weight}{Parameter to control which lambda is used to calculate
weights}

\item{seed}{seed}
}
\value{
\item{Ridge}{The model of ridge regression to calculate penalry.}
\item{AdaLasso}{The model of Adaptive Lasso for DFC extraction.}
\item{weights}{A data frame of extracted features and the weights. When
return_Model set FALSE, only weigths are returned.}
}
\description{
Extract DFC with Adaptive Lasso
}
