% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc.R
\name{dfc.matrix}
\alias{dfc.matrix}
\title{Extract DFC}
\usage{
\method{dfc}{matrix}(
  data,
  target_clusters,
  cluster_label = NULL,
  gamma = 1,
  seed = NULL,
  lambda_penalty = "1se",
  lambda_weight = "1se",
  SIS = TRUE,
  min_feature = NULL,
  max_feature = NULL,
  ...
)
}
\arguments{
\item{data}{scaled count matrix}

\item{target_clusters}{The target cluster number}

\item{cluster_label}{Logical or binary vector labeling target clusters}

\item{gamma}{Parameter to control the effect of penalty}

\item{seed}{seed}

\item{lambda_penalty}{Parameter to control which lambda is used to calculate
penalty, "min" or "1se". "min" : binomial deviance is minimized.
"1se" : the number of extracted features are minimized within error range of
"min".}

\item{lambda_weight}{Parameter to control which lambda is used to calculate
weights.}

\item{SIS}{Perform screening by SIS or not}

\item{min_feature}{minimum number to extract}

\item{max_feature}{maximum number to extract}

\item{...}{...}
}
\value{
\item{Ridge}{The model of ridge regression to calculate penalry.}
\item{AdaLasso}{The model of Adaptive Lasso for DFC extraction.}
\item{weights}{A data frame of extracted features and the weights. When
return_Model set FALSE, only weigths are returned.}
}
\description{
Extract DFC
}
