% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dfc}
\alias{dfc}
\title{Extract DFC}
\usage{
dfc(data, ...)
}
\arguments{
\item{data}{SeuratObject or gene x cell Matrix}

\item{...}{...}
}
\value{
\item{Ridge}{The model of ridge regression to calculate penalry.}
\item{AdaLasso}{The model of Adaptive Lasso for DFC extraction.}
\item{weights}{A data frame of extracted features and the weights. When
return_Model set FALSE, only weigths are returned.}
}
\description{
Extract DFC
}
