% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc.R
\name{dfc.Seurat}
\alias{dfc.Seurat}
\title{Extract DFC}
\usage{
\method{dfc}{Seurat}(
  data,
  target_clusters,
  assay = NULL,
  gamma = 1,
  return_Model = FALSE,
  seed = NULL,
  lambda_penalty = "1se",
  lambda_weight = "1se",
  SIS = TRUE,
  min_feature = NULL,
  max_feature = NULL
)
}
\arguments{
\item{data}{SeuratObject}

\item{target_clusters}{The target cluster number}

\item{assay}{The assay to use. In default, the return of DefaultAssay() is
used}

\item{gamma}{The parameter to control the effect of penalty}

\item{return_Model}{Return regression models or only weights}

\item{seed}{seed}

\item{lambda_penalty}{Parameter to control which lambda is used to calculate
penalty, "min" or "1se". "min" : binomial deviance is minimized.
"1se" : the number of extracted features are minimized within error range of
"min".}

\item{lambda_weight}{Parameter to control which lambda is used to calculate
weights.}

\item{SIS}{Perform screening by SIS or not}

\item{min_feature}{minimum number to extract}

\item{max_feature}{maximum number to extract}
}
\value{
\item{Ridge}{The model of ridge regression to calculate penalry.}
\item{AdaLasso}{The model of Adaptive Lasso for DFC extraction.}
\item{weights}{A data frame of extracted features and the weights. When
return_Model set FALSE, only weigths are returned.}
}
\description{
Extract DFC
}
