% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc_classify.R
\name{dfc_classify.dgCMatrix}
\alias{dfc_classify.dgCMatrix}
\title{Specificity based DFC classification}
\usage{
\method{dfc_classify}{dgCMatrix}(data, dfc_res, ...)
}
\arguments{
\item{data}{The same one as when \code{dfc()} is run.}

\item{dfc_res}{A result of \code{dfc()}.}

\item{...}{...}
}
\value{
Result data frame with DFC class.
}
\description{
Specificity based DFC classification
}
