% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc_classify.R
\name{dfc_classify.matrix}
\alias{dfc_classify.matrix}
\title{Specificity based DFC classification}
\usage{
\method{dfc_classify}{matrix}(
  data,
  dfc_res,
  rate_threshold = 0.25,
  cluster_threshold = NULL,
  eval_method = "min",
  ...
)
}
\arguments{
\item{data}{The same one as when \code{dfc()} is run.}

\item{dfc_res}{A result of \code{dfc()}.}

\item{rate_threshold}{Parameter to judge which the feature is expressed or
not in the cluster.}

\item{cluster_threshold}{Parameter to control expression specificity of
strong feature.}

\item{...}{...}
}
\value{
Result data frame with DFC class.
}
\description{
Specificity based DFC classification
}
