% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc_classify.R
\name{dfc_classify.Seurat}
\alias{dfc_classify.Seurat}
\title{Specificity based DFC classification}
\usage{
\method{dfc_classify}{Seurat}(
  data,
  dfc_res,
  assay = NULL,
  rate_threshold = 0.25,
  cluster_threshold = NULL
)
}
\arguments{
\item{data}{The same one as when \code{dfc()} is run.}

\item{dfc_res}{A result of \code{dfc()}.}

\item{assay}{The assay to use. In default, the return of \code{DefaultAssay()} is
used}

\item{rate_threshold}{Parameter to judge which the feature is expressed or
not in the cluster.}

\item{cluster_threshold}{Parameter to control expression specificity of
strong feature.}
}
\value{
Result data frame with DFC class.
}
\description{
Specificity based DFC classification
}
