% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc.R
\name{sis}
\alias{sis}
\title{Sure independence screening for pre-screening of features.}
\usage{
sis(data, label, min_feature = NULL, max_feature = NULL)
}
\arguments{
\item{data}{scaled count gene x cell matrix}

\item{label}{logical or binary vector labeling target clusters}

\item{min_feature}{minimum number to extract}

\item{max_feature}{maximum number to extract}
}
\value{
screened data
}
\description{
Sure independence screening for pre-screening of features.
}
