% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc.R
\name{dfc.Seurat}
\alias{dfc.Seurat}
\title{Extract DFC}
\usage{
\method{dfc}{Seurat}(data, target_clusters, assay = NULL, ...)
}
\arguments{
\item{data}{SeuratObject}

\item{target_clusters}{The target cluster number}

\item{assay}{The assay to use. In default, the return of DefaultAssay() is
used.}

\item{...}{...}
}
\value{
\item{Ridge}{The model of ridge regression to calculate penalry.}
\item{AdaLasso}{The model of Adaptive Lasso for DFC extraction.}
\item{weights}{A data frame of extracted features and the weights. When
return_Model set FALSE, only weigths are returned.}
}
\description{
Extract DFC
}
