% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc.R
\name{dfc_classify.dgCMatrix}
\alias{dfc_classify.dgCMatrix}
\title{Specificity based DFC classification}
\usage{
\method{dfc_classify}{dgCMatrix}(
  dfc_res,
  data,
  cluster_label,
  rate_threshold = 0.25,
  cluster_threshold = NULL
)
}
\arguments{
\item{dfc_res}{A result of \code{dfc()}.}

\item{data}{The same one as when \code{dfc()} is run.}

\item{cluster_label}{A vector denoting the clusters of each cell
corresponding to the columns of data.}

\item{rate_threshold}{Parameter to judge which the feature is expressed or
not in the cluster.}

\item{cluster_threshold}{Parameter to control expression specificity of
strong feature.}
}
\value{
Result data frame with DFC class.
}
\description{
Specificity based DFC classification
}
