% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dfc_classify}
\alias{dfc_classify}
\title{Specificity based DFC classification}
\usage{
dfc_classify(data, ...)
}
\arguments{
\item{data}{The same one as when \code{dfc()} is run.'}

\item{...}{...}
}
\value{
Result data frame with DFC class.
}
\description{
Specificity based DFC classification
}
